+++ packages/i18n/index.ts
@@ -1,3 +1,0 @@
-export * from './src/StreamI18n';
-export * from './src/types';
-export * from './src/utils';