+++ packages/react-native-sdk/expo-config-plugin/fixtures/index.ts
@@ -1,17 +1,18 @@
 import path from 'path';
 import fs from 'fs';
 
-type FileNames =
+type FileName =
   | 'AppDelegate.mm'
   | 'MainActivity.java'
   | 'MainApplication.java'
-  | 'AndroidManifest.xml';
+  | 'AndroidManifest.xml'
+  | 'app-build.gradle';
 
-export function getFixture(name: FileNames): string {
+export function getFixture(name: FileName): string {
   const filepath = path.join(__dirname, name);
   return fs.readFileSync(filepath, 'utf8');
 }
 
-export function getFixturePath(name: FileNames): string {
+export function getFixturePath(name: FileName): string {
   return path.join(__dirname, name);
 }