+++ packages/react-native-sdk/src/icons/LeaveStreamIcon.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { Path, Svg } from 'react-native-svg';
+
+export const LeaveStreamIcon = () => {
+  return (
+    <Svg viewBox="0 0 25 24" fill="none">
+      <Path
+        d="M18.75 5.75L16.9875 7.5125L20.2125 10.75H7.5V13.25H20.2125L16.9875 16.475L18.75 18.25L25 12L18.75 5.75ZM2.5 3.25H12.5V0.75H2.5C1.125 0.75 0 1.875 0 3.25V20.75C0 22.125 1.125 23.25 2.5 23.25H12.5V20.75H2.5V3.25Z"
+        fill="#FCFCFC"
+      />
+    </Svg>
+  );
+};