+++ packages/react-native-sdk/src/icons/index.tsx
@@ -14,3 +14,8 @@ export * from './Spotlight';
 export * from './ScreenShare';
 export * from './Reaction';
 export * from './TopViewBackground';
+export * from './Eye';
+export * from './ShieldBadge';
+export * from './StartStreamIcon';
+export * from './EndStreamIcon';
+export * from './LeaveStreamIcon';