+++ sample-apps/react/react-video-demo/src/components/Icons/index.ts
@@ -18,6 +18,7 @@ export { FullScreen } from './FullScreen';
 export { Mic } from './Mic';
 export { MicMuted } from './MicMuted';
 export { Latency } from './Latency';
+export { Layout } from './Layout';
 export { Leave } from './Leave';
 export { Like } from './Like';
 export { Loading } from './Loading';